+++ stepup/queue/actions.py
@@ -19,8 +19,10 @@
 # --
 """StepUp Queue package."""
 
+import argparse
 import contextlib
 import os
+import shlex
 
 from path import Path
 
@@ -32,13 +34,18 @@ from .sbatch import InpDigestError, submit_once_and_wait
 
 
 def sbatch(argstr: str, work_thread: WorkThread) -> int:
-    ext = ".sh" if argstr == "" else argstr
+    # Use argparse to parse the argstr
+    parser = argparse.ArgumentParser()
+    parser.add_argument("ext", nargs="?", default=".sh")
+    parser.add_argument("--rc", default=None)
+    args = parser.parse_args(shlex.split(argstr))
+
     if string_to_bool(os.getenv("STEPUP_QUEUE_RESUBMIT_CHANGED_INPUTS", "0")):
         with contextlib.suppress(InpDigestError):
-            return submit_once_and_wait(work_thread, ext)
+            return submit_once_and_wait(work_thread, args.ext, args.rc)
         # Cancel running job (if any), clean log and resubmit
         path_log = Path("slurmjob.log")
         job_id, cluster = read_jobid_cluster(path_log)
         work_thread.runsh(f"scancel -M {cluster} {job_id}")
         path_log.remove_p()
-    return submit_once_and_wait(work_thread, ext)
+    return submit_once_and_wait(work_thread, args.ext, args.rc)